refactor: reorganize chat UI modules
