+++ apps/web/src/ui/chat/shell/layout/ChatLayoutProvider.tsx
