+++ apps/web/src/ui/chat/shell/panel/ChatPanel.module.css
