+++ apps/web/src/ui/chat/shell/panel/ChatPanelHeader.tsx
