+++ apps/web/src/ui/chat/shell/panel/FileAttachment.tsx
