+++ apps/web/src/ui/chat/stream/display/chatTranscriptMarkdown.ts
