+++ apps/web/src/ui/chat/stream/display/thinkingSummary.ts
