+++ apps/web/src/ui/chat/stream/display/toolCallDisplay.ts
