+++ apps/web/src/ui/chat/stream/hooks/chatAutoScroll.ts
