+++ apps/web/src/ui/chat/stream/hooks/useDesktopEnterToSend.ts
